defconfig: replace broken CONFIG_SUSPEND_TIME with CONFIG_PM_DEBUG
* kernel/power/suspend_time.c is an Android addition (even absent
from Google's kernels 4.4 and higher) that relies on the
read_persistent_clock function to work around the fact that the system
clock (software) does not increment during suspend.
* Recent msm platforms (including msm8916/msm8976) use the standard
CONFIG_ARM_ARCH_TIMER, where the read_persistent_clock is stubbed out
and returns zero via a call to dummy_clock_access. This results
in many confusing messages that look like this:

    Suspended for 0.000 seconds

when it is in fact only the timekeeping that is off.
* On the other hand, CONFIG_PM_DEBUG is even enabled in many defconfigs
and in the mainline kernel/configs/android-recommended.config file.
* Since both config options provide mostly the same information, we are
keeping the one with less dependencies.

Change-Id: If5e87623f2484320cd475a9f270e1b040a1f7a4e
Signed-off-by: Vladimir Oltean <olteanv@gmail.com>